crud+paging+화면

+++ src/main/java/com/example/demo/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.demo.exception;
 
 public class PostNotFoundException extends RuntimeException {
-    public PostNotFoundException(Long poNum) {
+    public PostNotFoundException(Integer poNum) {
         super("Could not find post with poNum " + poNum);
     }
 }
